Add files via upload
binary search tree there is two nodes one is left child and other is a right child and a value. we will insert values in the binary tree using insert and print it  using the three traversals or the print function. Deletion there are four case leaf node, two children, one child and leaf node. and last there is a search in which we search value in bst.